requirements.txt
aiohttp>=3.9.0
pandas>=2.0.0
numpy>=1.26.0
ta>=0.11.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-multipart>=0.0.9
